add plugin architecture
This commit introduces an experimental mechanism to extend and enhancing
the IDE functionalities via plugins. A plugin is a way for a third party
to extend the functionality of an application.

GramLab plugins are build on [PF4J](https://github.com/decebals/pf4j),
an open source and lightweight plugin framework for java, with minimal
dependencies and easily extensible. Plugins are distributed in ZIP files
which bundled all runtime dependences, and that can be installed without
difficulty copying them into the `App/plugins` folder.

To develop a new plugin the [GramLab Skeleton plugin](core-plugins/skel)
will give a starting point.