add a new feature to process .info files as markdown documents
As illustrated on the Fig. 2.14 "Parameterizing the application of
dictionaries" of the User's Manual, a dictionary `foo.dic` could be
documented creating a `foo.txt` file. Commit f0fd037 extend this
behavior to `foo.info` files.

This commit introduces an early feature to process a `foo.info` file
as a markdown document using [Txtmark](https://github.com/rjeschke/txtmark),
a Java markdown processor.