feature #16: Allow to create a non-existent sub-graph
Currently, when a user creates a box in a graph with a non-existing
sub-graph e.g. ``:foo``, it turns red. When the user clicks on the
red box and selects ``Open subgraph`` a error dialog window stating
``Cannot find foo`` appears. 

This commit replaces the ``Cannot find foo`` dialog box with a dialog
asking the user to ``create a new one on /some/path/foo.grf``. If the
answer is ``Yes``, the IDE opens an empty graph named ``foo.grf``
on ``/some/path``. Any parent directory of ``/some/path`` will be
also created if it didn't exist yet.

Close #16