[feature #24] Add Find and replace for graphs
Currently, there is no way to find and replace the content of one or
more boxes in the graph editor. This commit implements such a feature.

To access the Find and Replace dialog box, from the FSGraph menu,
select Find and Replace or click the magnifying glass icon on the graph
editor Toolbar.

Resolve #23