[exp enhance] Draw input variable names above parenthesis
It is possible to select parts of a text sequence recognized by a
grammar using input variables (User's Manual, section 5.2.5). It
is also possible to capture portions of the outputs produced by
a grammar using output variables (User's Manual, section 6.8)

Currently, input and outputs variables are only differenced by a
color: red for inputs, blue for outputs. However, this has a major
drawback that if the graph is exported and converted to black and
white or grayscale colors, there is no longer any way to
difference the type of a variable.

This commit implements an experimental change in the way that
variables are drawn:

- input variables:  their name above the parenthesis, red color
- output variables: their name below the parenthesis, blue color

Whereas color conventions are maintained, from now on, the name
of the input variables is drawn above the parenthesis. This is
consistent with the convention input-above / output-below already
used for input/output labels and resolves the ambiguity to define
the kind of variable on monochrome or grayscale images.