[fix #77] Resolve NullPointerException in ExportTextAsPOSListDialog This commit resolves a null pointer exception occuring when a user tries to export an automaton as a POS list on some corpora.
When ExportTextAsPOSListDialog is unable to load the graph corresponding to a sentence, it ignores it and processes the remaining sentences.

Once the exportation is over, a warning message will be displayed in order to notify the user about which sentence has been skipped (if there is none then no message is displayed).