[enhance] Improve korean graph paths exploring
This commit allows to print a readable output while 
exploring Korean graph paths.  It fixes #131 and it is 
associated with the issue unitex-core/87

It also adds an argument "-D name" to give the list
of morphological dictionaries specified by the user in
Preferences. It fixes #132 and it is associated with 
the issue unitex-core/91

Resolves #131, #132 
See also #135

Co-authored-by: Eradan94 <44415213+Eradan94@users.noreply.github.com>
Co-authored-by: BIGUENET DENIS <dbiguene@localhost.localdomain>